+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added a note in README about os support and the controller requiremenet
 - Say `uploading` immediately after pressing the F4 keybinding
 - Add save/load menu tts
+- Added text beeing toggle
  
 ## [0.4.6]
 ### Added